+++ packages/shared/common/src/internal/events/index.ts
@@ -5,6 +5,7 @@ import InputEvalEvent from './InputEvalEvent';
 import InputEvent from './InputEvent';
 import InputIdentifyEvent from './InputIdentifyEvent';
 import InputMigrationEvent from './InputMigrationEvent';
+import type { LDInternalOptions } from './LDInternalOptions';
 import NullEventProcessor from './NullEventProcessor';
 import shouldSample from './sampling';
 
@@ -18,4 +19,5 @@ export {
   EventProcessor,
   shouldSample,
   NullEventProcessor,
+  LDInternalOptions,
 };