+++ packages/shared/sdk-server-edge/src/api/LDClient.ts
@@ -1,6 +1,6 @@
 import { EventEmitter } from 'node:events';
 
-import { Info, LDClientImpl, LDOptions } from '@launchdarkly/js-server-sdk-common';
+import { Info, internal, LDClientImpl, LDOptions } from '@launchdarkly/js-server-sdk-common';
 
 import EdgePlatform from '../platform';
 import createCallbacks from './createCallbacks';
@@ -9,16 +9,19 @@ import createOptions from './createOptions';
 /**
  * The LaunchDarkly SDK edge client object.
  */
-export class LDClient extends LDClientImpl {
+export default class LDClient extends LDClientImpl {
   emitter: EventEmitter;
 
-  // sdkKey is only used to query featureStore, not to initialize with LD servers
-  constructor(sdkKey: string, platformInfo: Info, options: LDOptions) {
+  // clientSideID is only used to query the edge key-value store and send analytics, not to initialize with LD servers
+  constructor(clientSideID: string, platformInfo: Info, options: LDOptions) {
     const em = new EventEmitter();
     const platform = new EdgePlatform(platformInfo);
-    super('n/a', platform, createOptions(options), createCallbacks(em));
+    const internalOptions: internal.LDInternalOptions = {
+      analyticsEventPath: `/events/bulk/${clientSideID}`,
+      diagnosticEventPath: `/events/diagnostic/${clientSideID}`,
+      includeAuthorizationHeader: false,
+    };
+    super(clientSideID, platform, createOptions(options), createCallbacks(em), internalOptions);
     this.emitter = em;
   }
 }
-
-export default LDClient;